+++ package.json
@@ -9,8 +9,10 @@
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.5",
     "firebase": "^11.6.0",
+    "particles.js": "^2.0.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.5.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
