+++ src/assets/images/teardrop-homepage.png
