+++ src/assets/styles/scss/_display.scss
@@ -5,4 +5,7 @@
 
 .particles-container {
     height: 100vh;
+    width: 100%;
+    border-radius: 5px;
+    padding: 1rem 0;
 }
