+++ src/assets/styles/scss/style.scss
@@ -3,4 +3,6 @@
 @import "fonts";
 @import "~bootstrap/scss/bootstrap";
 
-@import "form";
+@import "homepage";
+@import "control";
+@import "display";
