+++ src/index.js
@@ -1,18 +1,21 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './assets/styles/scss/style.scss';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import Homepage from './components/Homepage.tsx';
+import ControlForm from './components/ControlForm.tsx';
+import Display from './components/Display.tsx';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+    <React.StrictMode>
+        <BrowserRouter>
+            <Routes>
+                <Route path="/" element={<Homepage />} />
+                <Route path="/control" element={<ControlForm />} />
+                <Route path="/display" element={<Display />} />
+            </Routes>
+        </BrowserRouter>
+    </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
