--- teardrop/src/assets/styles/scss/_display.scss
@@ -1,11 +1,0 @@
-// Particles Display
-.display-container {
-    max-width: none;
-}
-
-.particles-container {
-    height: 100vh;
-    width: 100%;
-    border-radius: 5px;
-    padding: 1rem 0;
-}
