--- teardrop/src/assets/styles/scss/_homepage.scss
@@ -1,44 +1,0 @@
-// Homepage
-.heading,
-.homepage-content strong,
-.link-button {
-    font-weight: bold;
-    font-style: italic;
-    text-transform: uppercase;
-}
-
-.heading {
-    font-size: 3.5rem;
-    letter-spacing: -10%;
-}
-
-.homepage-content {
-    overflow-x: hidden;
-
-    p {
-        font-size: 1.5rem;
-    }
-
-    .link-button {
-        border: 2px solid $accent-green;
-        color: $white;
-        border-radius: 5px !important;
-
-        &:hover {
-            background-color: $accent-green;
-            color: $body-bg;
-        }
-    }
-}
-
-@media (max-width: 576px) {
-    .heading {
-        font-size: 2rem;
-    }
-
-    .homepage-content {
-        p {
-            font-size: 1.2rem;
-        }
-    }
-}
