--- teardrop/src/assets/styles/scss/_layout.scss
@@ -1,30 +1,0 @@
-// Layout
-.double-border {
-    border: 5px solid $indigo;
-    padding: 10px;
-    position: relative;
-    border-radius: 20px;
-    max-width: 800px;
-    box-sizing: border-box;
-
-    &::before {
-        content: '';
-        position: absolute;
-        top: -25px;
-        left: 18px;
-        right: 18px;
-        bottom: -25px;
-        border: 5px solid $indigo;
-        border-radius: 20px;
-        pointer-events: none;
-    }
-}
-
-@media (max-width: 576px) {
-    .double-border {
-        overflow-x: hidden;
-        &::before {
-            content: none;
-        }
-    }
-}
