--- teardrop/src/assets/styles/scss/style.scss
@@ -1,8 +1,0 @@
-@import "variables";
-@import "layout";
-@import "fonts";
-@import "~bootstrap/scss/bootstrap";
-
-@import "homepage";
-@import "control";
-@import "display";
